requirements.txt
pyqt6-sip
# Secure Vault
# Python dependencies

PyQt6>=6.6.0
pynacl>=1.5.0
